fix argv
