+++ src/cli/index.ts
@@ -1,8 +1,8 @@
-
-import program, { Command } from 'commander';
-import { CommandError } from '../error';
-import { i18n } from '../utils';
+import program, {Command} from 'commander';
+import {CommandError} from '../error';
+import {i18n} from '../utils';
 import CliManager from './cli-manager';
+
 const description = `${i18n.__('Directly use serverless devs to use components, develop and manage applications without yaml configuration.')}
 
     ${i18n.__('Example:')}
@@ -12,13 +12,14 @@ const description = `${i18n.__('Directly use serverless devs to use components, 
     `;
 
 
-program
+const cliCommand = program
     .name('s cli')
     .usage('s cli [component] [method] [options]')
     .option('-a, --access [access-alias]', i18n.__('Specify the key name'))
     .option('-p, --props [json-string]', i18n.__('The json string of props'))
     .helpOption('-h, --help', i18n.__('Display help for command'))
-    .description(description).addHelpCommand(false).parse(process.argv);
+    .description(description).addHelpCommand(false);
+
 const subCommandName = process.argv[2];
 if (subCommandName) {
     const execCommand = new Command(subCommandName);
@@ -29,13 +30,36 @@ if (subCommandName) {
 }
 
 (async () => {
-    if (program.args.length === 0) {
+
+    if ((process.argv.length == 2) || (process.argv.length == 3 && ['-h', '--help'].includes(process.argv[2]))) {
         program.help();
+    } else {
+        const tempCommand = process.argv[3]
+        let start = false;
+        const processArgv: string[] = [];
+        let params: string[] = [];
+        for (let i = 0; i < process.argv.length; i++) {
+            if (!start) {
+                processArgv.push(process.argv[i]);
+            } else {
+                params.push(process.argv[i]);
+            }
+            if (process.argv[i] === tempCommand) {
+                start = true;
+            }
+        }
+        if (params.length !== 0) {
+            process.env.temp_params = params.concat(process.env.temp_params).join(' ');
+        }
+        processArgv.push(tempCommand)
+        process.argv = processArgv;
+        cliCommand.parse(process.argv)
+        const [component, command] = program.args;
+        const {access, props} = program as any;
+        const cliManager = new CliManager({command, component, access, props});
+        cliManager.init();
+
     }
-    const [component, command] = program.args;
-    const { access, props } = program as any;
-    const cliManager = new CliManager({ command, component, access, props });
-    cliManager.init();
 })().catch(err => {
     throw new CommandError(err.message);
 });
